requirements.txt
langchain-core
langchain-openai
langchain-ollama
pydantic
PyYAML
python-dotenv